fixes after rebase
